build and install macOS, win same as Linux (#28)
The wxPython CI scripts build macOS and Windows like we're building Linux
If we do the same, this will make the same artifacts available on both platforms
(i.e. will provide missing wxPython headers [and wx-config on macOS])

--no_magic on Linux is what was including the wxWidget headers and putting
the libraries in 'lib/' instead of Python site-packages. Use it on
macOS and Windows