--use_syswx when building wxpython
with only --no_magic for build_py, macOS has hardcoded link paths
to the wxwidgets build directory in all of the python extension module .so's
--use_syswx will cause build.py to lookup the wxwidgets we just installed in
and then the rpath fixing done by conda-build makes everythign work.

Windows requires a patch to build.py to allow use of --no_magic.  I am
going to follow up with wxpython community in
https://discuss.wxpython.org/t/including-wxwidgets-headers-in-windows-and-macos-builds/34471/2
regarding upstreaming this change. Also switched Windows to use_syswx so that it matches
the other platform builds as much as possible.