4.2.0 still uses attrdict
The commit: https://github.com/wxWidgets/Phoenix/commit/2e9169effa9abf14f34f8436a791b8829eea7774
later changes this dependency (on attrdict3) to only be on
windows, but on 4.2.0 this is universally imported. Note
that attrdict3 appears to be a separate project to attrdict and
only the latter is available on conda-forge - since this is
a build-time only dependency I'm hoping that it's similar enough.